+++ CHANGELOG.md
@@ -21,7 +21,8 @@ All Notable changes to `bakame/http-strucured-fields` will be documented in this
 
 - Improve annotation using `@phpstan-type`
 - `Value` internal class to improve Item public API;
-- **[BC Break]**  `::fromAssociative` and `::fromPair` the `$parameters` argument is now required;
+- **[BC Break]** `::fromAssociative` and `::fromPair` the `$parameters` argument is now required;
+- **[BC Break]** `MemberOrderedMap` instances can no longer be added to `Dictionary` or `OuterList` instances.
 
 ### Deprecated